Fixed Total Amount Decreased

+++ modules/Delete/BudgetDelete.js
@@ -11,6 +11,9 @@ async function deleteBudget(UserID, requestInfo) {
     const { itemsID } = requestInfo;
     const itemsAmount = -(await db.getBudgetAmount(UserID, itemsID));
 
+    if (!itemsAmount) {
+        throw new ErrorHandler("Unexisting Data", "Data does not exist in the database", 404);
+    }
     const status = await db.deleteUserBudget(UserID, itemsID, itemsAmount);
     const serverResponse = new Response('Data SuccessFully Deleted', 200, status);
     return serverResponse;
